format, cleanup logs

+++ src/interface/CErc20Interface.sol
@@ -1,7 +1,5 @@
 pragma solidity 0.8.25;
 
 interface CErc20Interface {
-
-    function mint(uint mintAmount) external returns (uint);
-
+    function mint(uint256 mintAmount) external returns (uint256);
 }
